% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{consecutive}
\alias{consecutive}
\title{Find consective elements}
\usage{
consecutive(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Logical vector indicating non-consecutive breaks
}
\description{
Given a numeric vector, returns a logical vector indicating whether
each vector element is consecutive with the previous element.
}
\examples{
consecutive(c(1900, 1901, 1903, 1904))

}
