% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.hmm.R
\docType{package}
\name{fc.hmm-package}
\alias{fc.hmm}
\alias{fc.hmm-package}
\title{fc.hmm}
\description{
Utility functions used in the R scripts for this project placed
into a separate package for isolation and testing.
}
\author{
\strong{Maintainer}: Joshua Krusell \email{joshua.krusell@gu.se}

}
\keyword{internal}
