% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group_breaks}
\alias{group_breaks}
\title{Grouping indices by boolean breaks}
\usage{
group_breaks(x)
}
\arguments{
\item{x}{Logical vector indicating breaks}
}
\value{
Integer vector of grouping indices
}
\description{
Given a logical vector, return a vector of the same length with the
indices for each group of of consecutive FALSE values terminated by
the first TRUE value.
}
\details{
This is used to identify conflict episodes given a vector
indicating whether a conflict-year has experienced a termination.
}
\examples{
group_breaks(c(T, F, F, T, F))

}
