% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{min_max}
\alias{min_max}
\title{Min-max normalization}
\usage{
min_max(x)
}
\arguments{
\item{x}{Numeric vector to re-scale}
}
\value{
Re-scaled numeric vector
}
\description{
This function applies min-max normalization to constrain a numeric
vector between \link{0, 1}.
}
\details{
The transformation is defined as \eqn{\frac{(x - min(x))}{(max(x) - min(x))}}
}
\seealso{
\code{\link[=normalize]{normalize()}}
}
