% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize}
\alias{normalize}
\title{Standard score normalization}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric vector to re-scale}
}
\description{
Convenience wrapper to \code{scale} that returns a numeric vector after
applying standard score normalization.
}
