% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{polynomial}
\alias{polynomial}
\title{Polynomial transformation}
\usage{
polynomial(df, var, n)
}
\arguments{
\item{df}{Data frame}

\item{var}{Character name of the variable to transform}

\item{n}{Integer number of polynomial terms}
}
\value{
A data frame with \code{n} additional columns.
}
\description{
Convenience wrapper to the \code{poly} function that adds \code{n} polynomial
terms to a data frame for variable \code{var}. Note, this function does
not use non-standard evaluation so the variable name must be
quoted.
}
