% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{posterior_transitions}
\alias{posterior_transitions}
\title{Predicted transition probabilities}
\usage{
posterior_transitions(fit, ...)
}
\arguments{
\item{fit}{A \code{cmdstanr} object containing the posterior draws}

\item{X}{Covariate matrix representing \code{nrow(X)} observations with
\code{ncol(X)} matching columns to the original model run data}

\item{unit_id}{Numeric vector of conflict episode IDs from the
original model run data}

\item{tiv}{Numeric vector of TIV values}

\item{from}{Integer representing the latent state to transition from}

\item{ndraws}{Number of posterior draws to use in the computation}

\item{mc.cores}{Number of cores to be used by OpenMP}
}
\value{
A \code{ndraw} length list of matrices with dimensions \code{K} x
\code{length(tiv)}, \emph{i.e.}, the average posterior probability for
transition to each state given the different values of TIV for
each posterior draw from \code{fit}.
}
\description{
This function computes the average predicted transition probability
from state \code{from} to each state 1, \ldots, K for each value of
\code{tiv} across all transition cases in \code{X}.
}
\note{
Warning, this function assumes that the first column of \code{X},
and correspondingly the first row for each array element in
\code{beta}, corresponds to the TIV variable. This is a bad
practice; however, this function is not meant to be a general
library function for other projects and our assumption
simplifies the code across a number of different scripts.
}
