% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{roll_mean}
\alias{roll_mean}
\title{Rolling mean}
\usage{
roll_mean(x, n)
}
\arguments{
\item{x}{Numeric vector for rolling mean}

\item{n}{Window size}
}
\description{
Calculate the left-inclusive rolling mean with a \code{n} size window.
}
\examples{
roll_mean(1:10, 3)

}
\seealso{
\code{\link[=window]{window()}}
}
