% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{window}
\alias{window}
\title{Windowing function}
\usage{
window(x, n)
}
\arguments{
\item{x}{Vector}

\item{n}{Window size}
}
\value{
A list of size \code{length(x)} where each element is an \code{n}
length vector.
}
\description{
Returns the list of moving windows of size \code{n} for a vector. For
each element in \code{x}, a window is defined as a vector of the
previous \code{(n-1)} elements and the current element.
}
\details{
When there are fewer than (n-1) elements to the left of the current
element in \code{x}, the window is padded with \code{NA}.
}
\examples{
window(1:10, 5)

}
